+++ pytket/qir/conversion/conversion.py
@@ -47,6 +47,7 @@ from pytket.circuit.logic_exp import (
     BitOr,
     BitWiseOp,
     BitXor,
+    BitZero,
     RegAdd,
     RegAnd,
     RegEq,
@@ -98,9 +99,10 @@ _TK_CLOPS_TO_PYQIR_BIT: dict = {
     BitEq: lambda b: partial(b.icmp, IntPredicate.EQ),
 }
 
-_TK_CLOPS_TO_PYQIR_BIT_NO_PARAM: list = [
-    BitOne,
-]
+_TK_CLOPS_TO_PYQIR_BIT_NO_PARAM: dict = {
+    BitOne: 1,
+    BitZero: 0,
+}
 
 
 class QirGenerator:
@@ -560,7 +562,7 @@ class QirGenerator:
 
             return output_instruction  # type: ignore
         else:
-            raise ValueError(f"unsupported bisewise operation {type(bit)}")
+            raise ValueError(f"unsupported bitwise operation {type(bit)}")
 
     def get_wasm_sar(self) -> dict[str, str]:
         return self.wasm_sar_dict
@@ -905,7 +907,12 @@ class QirGenerator:
 
                 elif type(op.get_exp()) in _TK_CLOPS_TO_PYQIR_BIT_NO_PARAM:
                     # classical ops without parameters
-                    output_instruction = pyqir.const(self.qir_int_type, 1)
+                    output_instruction = pyqir.const(
+                        self.qir_bool_type,
+                        _TK_CLOPS_TO_PYQIR_BIT_NO_PARAM[type(op.get_exp())],
+                    )
+                    returntypebool = True
+                    result_index = command.args[-1].index[0]
 
                 elif type(op.get_exp()) in _TK_CLOPS_TO_PYQIR_BIT:
                     # classical ops acting on bits returning bit
@@ -920,7 +927,7 @@ class QirGenerator:
 
                     # add function to module
                     returntypebool = True
-                    result_index = command.args[-1].index[0]  # todo
+                    result_index = command.args[-1].index[0]
                     output_instruction = _TK_CLOPS_TO_PYQIR_BIT[type(op.get_exp())](
                         module.builder
                     )(ssa_left, ssa_right)